legacy: addition of sql compatibility functions
* Adds new functions to permit the sql compatibility for MySQL and
  PostgreSQL. (addresses #3214)

+++ invenio/testsuite/test_legacy_dbquery.py
@@ -33,29 +33,31 @@ class TableUpdateTimesTest(InvenioTestCase):
 
     def _check_table_update_time(self, tablename):
         """Helper function to check update time of TABLENAME."""
-        from invenio.base.globals import cfg
-        # detect MySQL version number:
-        res = dbquery.run_sql("SELECT VERSION()")
-        mysql_server_version = res[0][0]
-        if mysql_server_version.startswith("5."):
-            # MySQL-5 provides INFORMATION_SCHEMA:
-            query = """SELECT UPDATE_TIME FROM INFORMATION_SCHEMA.TABLES
-                        WHERE table_name='%s' AND table_schema='%s' """ \
-                        % (tablename, cfg['CFG_DATABASE_NAME'])
-            tablename_update_time = str(dbquery.run_sql(query)[0][0])
-        elif mysql_server_version.startswith("4.1"):
-            # MySQL-4.1 has it on 12th position:
-            query = """SHOW TABLE STATUS LIKE "%s" """ % tablename
-            tablename_update_time = str(dbquery.run_sql(query)[0][12])
-        elif mysql_server_version.startswith("4.0"):
-            # MySQL-4.0 has it on 11th position:
-            query = """SHOW TABLE STATUS LIKE "%s" """ % tablename
-            tablename_update_time = str(dbquery.run_sql(query)[0][11])
-        else:
-            tablename_update_time = "MYSQL SERVER VERSION NOT DETECTED"
-        # compare it with the one detected by the function:
-        self.assertEqual(tablename_update_time,
-                         dbquery.get_table_update_time(tablename))
+        from invenio.ext.sqlalchemy import db
+        if db.engine.driver == 'mysql':
+            from invenio.base.globals import cfg
+            # detect MySQL version number:
+            res = dbquery.run_sql("SELECT VERSION()")
+            mysql_server_version = res[0][0]
+            if mysql_server_version.startswith("5."):
+                # MySQL-5 provides INFORMATION_SCHEMA:
+                query = """SELECT UPDATE_TIME FROM INFORMATION_SCHEMA.TABLES
+                            WHERE table_name='%s' AND table_schema='%s' """ \
+                            % (tablename, cfg['CFG_DATABASE_NAME'])
+                tablename_update_time = str(dbquery.run_sql(query)[0][0])
+            elif mysql_server_version.startswith("4.1"):
+                # MySQL-4.1 has it on 12th position:
+                query = """SHOW TABLE STATUS LIKE "%s" """ % tablename
+                tablename_update_time = str(dbquery.run_sql(query)[0][12])
+            elif mysql_server_version.startswith("4.0"):
+                # MySQL-4.0 has it on 11th position:
+                query = """SHOW TABLE STATUS LIKE "%s" """ % tablename
+                tablename_update_time = str(dbquery.run_sql(query)[0][11])
+            else:
+                tablename_update_time = "MYSQL SERVER VERSION NOT DETECTED"
+            # compare it with the one detected by the function:
+            self.assertEqual(tablename_update_time,
+                             dbquery.get_table_update_time(tablename))
 
     def test_single_table_update_time(self):
         """dbquery - single table (with indexes) update time detection."""
@@ -65,34 +67,39 @@ class TableUpdateTimesTest(InvenioTestCase):
 
     def test_empty_table_update_time(self):
         """dbquery - empty table (no indexes) update time detection."""
-        # NOTE: this tests unusual "None" branch of
-        # get_table_update_time()
-        # create empty test table
-        test_table = "tmpTESTTABLE123"
-        dbquery.run_sql("CREATE TABLE IF NOT EXISTS %s (a INT)" % test_table)
-        # run the test:
-        self._check_table_update_time(test_table)
-        # drop empty test table
-        dbquery.run_sql("DROP TABLE %s" % test_table)
+        from invenio.ext.sqlalchemy import db
+        if db.engine.driver == 'mysql':
+            # NOTE: this tests unusual "None" branch of
+            # get_table_update_time()
+            # create empty test table
+            test_table = "tmpTESTTABLE123"
+            dbquery.run_sql("CREATE TABLE IF NOT EXISTS %s (a INT)" %
+                            test_table)
+            # run the test:
+            self._check_table_update_time(test_table)
+            # drop empty test table
+            dbquery.run_sql("DROP TABLE %s" % test_table)
 
     def test_utf8_python_mysqldb_mysql_storage_chain(self):
         """dbquery - UTF-8 in Python<->MySQLdb<->MySQL storage chain."""
-        # NOTE: This test test creates, uses and destroys a temporary
-        # table called "test__invenio__utf8".
-        beta_in_utf8 = "β"  # Greek beta in UTF-8 is 0xCEB2
-        dbquery.run_sql(
-            "CREATE TEMPORARY TABLE test__invenio__utf8 "
-            "(x char(1), y varbinary(2)) DEFAULT CHARACTER SET utf8")
-        dbquery.run_sql(
-            "INSERT INTO test__invenio__utf8 (x, y) VALUES (%s, %s)",
-            (beta_in_utf8, beta_in_utf8))
-        res = dbquery.run_sql(
-            "SELECT x,y,HEX(x),HEX(y),LENGTH(x),LENGTH(y),"
-            "CHAR_LENGTH(x),CHAR_LENGTH(y) FROM test__invenio__utf8")
-        self.assertEqual(
-            res[0],
-            ('\xce\xb2', '\xce\xb2', 'CEB2', 'CEB2', 2L, 2L, 1L, 2L))
-        dbquery.run_sql("DROP TEMPORARY TABLE test__invenio__utf8")
+        from invenio.ext.sqlalchemy import db
+        if db.engine.driver == 'mysql':
+            # NOTE: This test test creates, uses and destroys a temporary
+            # table called "test__invenio__utf8".
+            beta_in_utf8 = "β"  # Greek beta in UTF-8 is 0xCEB2
+            dbquery.run_sql(
+                "CREATE TEMPORARY TABLE test__invenio__utf8 "
+                "(x char(1), y varbinary(2)) DEFAULT CHARACTER SET utf8")
+            dbquery.run_sql(
+                "INSERT INTO test__invenio__utf8 (x, y) VALUES (%s, %s)",
+                (beta_in_utf8, beta_in_utf8))
+            res = dbquery.run_sql(
+                "SELECT x,y,HEX(x),HEX(y),LENGTH(x),LENGTH(y),"
+                "CHAR_LENGTH(x),CHAR_LENGTH(y) FROM test__invenio__utf8")
+            self.assertEqual(
+                res[0],
+                ('\xce\xb2', '\xce\xb2', 'CEB2', 'CEB2', 2L, 2L, 1L, 2L))
+            dbquery.run_sql("DROP TEMPORARY TABLE test__invenio__utf8")
 
 
 class WashTableColumnNameTest(InvenioTestCase):
